Add Board#allowed? To see if point accessible
In the physical Connect Four game, the pieces fall straight down,
occupying the lowest available space within the column. The disks cannot
suspend in mid-air without any contact from a bottom disk. Thus, this
commit's implementation ensures that the constraint adhered to by only
ever allowing a limited number of points to insert the disk into.

And in introducing this implementations with the accompanying specs, the
older specs' variables had to be adjusted to adhere to the
points_allowed rule beforehand.