Edit Board class by re-imagining player input
Instead of tediously prompting the player to enter both column and row
as input, rather accept the column. Then allow the program to simply
place the token at the lowest possible position in that column. Then
have the program keep record of whether that column has any space left,
and indicate if full to the player, thus requesting them to select
different column.

This enforces the typical mechanics of the real-life game where they
only choose the column, and then the token falls to the lowest possible
row.